[pkl-extensions] Replace PklValueSource with AbstractPklValueSource
PklValueSource returned PModule directly, which is not java.io.Serializable
and therefore incompatible with Gradle's configuration cache. Remove the
concrete class, pklFile() factory functions, and parsePkl() extension, all
of which surfaced PModule to callers. Introduce AbstractPklValueSource as a
generic abstract base (doObtain pattern) with explicit KDoc noting the
config-cache incompleteness, and refactor PklPathValueSource to extend it.
